Issue temp: Imrpove new packages formatting & text
This tries to improve the formatting by adding line breaks and updates the wording in a few cases.
It also removes the checkbox for meeting general requirements as it's not clear how to confirm that from the linked document. As the issue template requests/includes the information to determine that already, it's not really needed anyways.